move the endpoint to routers directory
