+++ app/main.py
@@ -10,7 +10,7 @@ from fastapi.templating import Jinja2Templates
 from app.database import models
 from app.database.models import Event
 from app.database.database import SessionLocal, engine, get_db
-from app.routers.event import add_event, check_validation
+from app.routers import event
 
 models.Base.metadata.create_all(bind=engine)
 
@@ -18,9 +18,9 @@ models.Base.metadata.create_all(bind=engine)
 app = FastAPI()
 
 app.mount("/static", StaticFiles(directory="app/static"), name="static")
-# its dosent found "static" folder
-
 templates = Jinja2Templates(directory="app/templates")
+app.include_router(event.router)
+
 
 
 @app.get("/")
@@ -45,34 +45,34 @@ def profile(request: Request):
     })
 
 
-@app.get("/profile/{user_id}/EditEvent", response_class=HTMLResponse)
-async def insert_info(request: Request) -> HTMLResponse:
-    """Get request and return an html File"""
-    return templates.TemplateResponse("editevent.html",{"request": request})
+# @app.get("/profile/{user_id}/EditEvent", response_class=HTMLResponse)
+# async def insert_info(request: Request) -> HTMLResponse:
+#     """Get request and return an html File"""
+#     return templates.TemplateResponse("editevent.html",{"request": request})
 
 
-@app.post("/profile/{user_id}/EditEvent") # this func is soupose to change with the PR of Ode and Efrat and it will be change
-def create_event(user_id: int, event_title: str = Form(None), location: Optional[str] = Form(None), from_date: Optional[datetime] = Form(...),
-                to_date: Optional[datetime] = Form(...), link_vc: str = Form(None), content: str = Form(None),
-                 db = Depends(get_db)) -> dict:
-    """ required args - title, from_date, to_date, user_id, the 'from_date' need to be early from the 'to_date'.
-    check validation for the value, insert the new data to DB 
-    if the prosess success return True arg the event item, otherwith return False and the error msg """
-    success = False
-    error_msg = ""
-    new_event = ""
-    if event_title is None:
-        event_title = "No Title"
-    try:
-        if check_validation(from_date, to_date):
-            event_value = {'title': event_title, "location": location, "start_date": from_date, "end_date": to_date, "vc_link":link_vc, "content": content, 
-                "owner_id": user_id}
-            new_event = add_event(event_value, db)
-            success = True
-        else:
-            error_msg = "Error, Your date is invalid"
-    except Exception as e:
-        error_msg = e
-    finally:
-        return {"success": success, "new_event": new_event, "error_msg": error_msg}
+# @app.post("/profile/{user_id}/EditEvent") # this func is soupose to change with the PR of Ode and Efrat and it will be change
+# def create_event(user_id: int, event_title: str = Form(None), location: Optional[str] = Form(None), from_date: Optional[datetime] = Form(...),
+#                 to_date: Optional[datetime] = Form(...), link_vc: str = Form(None), content: str = Form(None),
+#                  db = Depends(get_db)) -> dict:
+#     """ required args - title, from_date, to_date, user_id, the 'from_date' need to be early from the 'to_date'.
+#     check validation for the value, insert the new data to DB 
+#     if the prosess success return True arg the event item, otherwith return False and the error msg """
+#     success = False
+#     error_msg = ""
+#     new_event = ""
+#     if event_title is None:
+#         event_title = "No Title"
+#     try:
+#         if check_validation(from_date, to_date):
+#             event_value = {'title': event_title, "location": location, "start_date": from_date, "end_date": to_date, "vc_link":link_vc, "content": content, 
+#                 "owner_id": user_id}
+#             new_event = add_event(event_value, db)
+#             success = True
+#         else:
+#             error_msg = "Error, Your date is invalid"
+#     except Exception as e:
+#         error_msg = e
+#     finally:
+#         return {"success": success, "new_event": new_event, "error_msg": error_msg}
