+++ app/routers/event.py
@@ -1,4 +1,51 @@
+from typing import Optional
+from datetime import datetime
+
+
+from fastapi import APIRouter, Depends, Request, Form 
+from fastapi.responses import HTMLResponse
+from fastapi.staticfiles import StaticFiles
+from fastapi.templating import Jinja2Templates
+
 from app.database.models import Event
+from app.database.database import get_db
+
+router = APIRouter()
+templates = Jinja2Templates(directory="app/templates")
+
+
+
+@router.get("/profile/{user_id}/EditEvent", response_class=HTMLResponse)
+async def insert_info(request: Request) -> HTMLResponse:
+    """Get request and return an html File"""
+    return templates.TemplateResponse("editevent.html",{"request": request})
+
+
+@router.post("/profile/{user_id}/EditEvent") # this func is soupose to change with the PR of Ode and Efrat and it will be change
+def create_event(user_id: int, event_title: str = Form(None), location: Optional[str] = Form(None), from_date: Optional[datetime] = Form(...),
+                to_date: Optional[datetime] = Form(...), link_vc: str = Form(None), content: str = Form(None),
+                 db = Depends(get_db)) -> dict:
+    """ required args - title, from_date, to_date, user_id, the 'from_date' need to be early from the 'to_date'.
+    check validation for the value, insert the new data to DB 
+    if the prosess success return True arg the event item, otherwith return False and the error msg """
+    success = False
+    error_msg = ""
+    new_event = ""
+    if event_title is None:
+        event_title = "No Title"
+    try:
+        if check_validation(from_date, to_date):
+            event_value = {'title': event_title, "location": location, "start_date": from_date, "end_date": to_date, "vc_link":link_vc, "content": content, 
+                "owner_id": user_id}
+            new_event = add_event(event_value, db)
+            success = True
+        else:
+            error_msg = "Error, Your date is invalid"
+    except Exception as e:
+        error_msg = e
+    finally:
+        return {"success": success, "new_event": new_event, "error_msg": error_msg}
+
 
 def check_validation(start_time, end_time) -> bool:
     """Check if the start_date is smaller then the end_time"""
