+++ app/templates/profile.html
@@ -9,6 +9,7 @@
             <div class="col-3">
                 <!-- Profile card -->
                 <div class="card card-profile mb-3">
+                    <img class="card-img-top profile-image rounded-circle mx-auto mt-3 d-block" src="{{ url_for('static', path='/images/profile.png') }}" alt="Profile image">
                     <div class="card-body">
                         <h6 class="card-title text-center mb-1">{{ username }}</h6>
                         <p class="card-text text-center mt-0">
