+++ tests/conftest.py
@@ -8,17 +8,17 @@ def client():
     return TestClient(app)
 
 
-@pytest.fixture
-def event_test_data():
-    a = {'title': "Test Title",
-        "location": "Fake City",
-        "start_date": datetime.today() - timedelta(1),
-        "end_date": datetime.today(),
-        "vc_link": "https://fakevclink.com",
-        "content": "Any Words",
-        "owner_id": 123
-        }
-    return a 
+# @pytest.fixture
+# def event_test_data():
+#     a = {'title': "Test Title",
+#         "location": "Fake City",
+#         "start_date": datetime.today() - timedelta(1),
+#         "end_date": datetime.today(),
+#         "vc_link": "https://fakevclink.com",
+#         "content": "Any Words",
+#         "owner_id": 123
+#         }
+#     return a 
 
 
 @pytest.fixture
